pyproject.toml
[project]
name = "vertical-slice-graphql-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.13",
    "fastapi>=0.119.1",
    "httpx>=0.28.1",
    "pydantic>=2.8",
    "pytest>=8.0",
    "python-dotenv>=1.0",
    "sqlalchemy>=2.0",
    "strawberry-graphql>=0.211",
    "uvicorn>=0.30",
]
